+++ _updates/world/0002_world_init.sql
@@ -8,6 +8,7 @@
 -- Global Unique Identifier ranges used:
 --
 -- - Spirit Healers, Azeroth: 1-500
+-- - Elevators, Azeroth:      1-100
 
 -- Provide variables for various item stats ----------------------------------
 -- -> classes and subclasses -------------------------------------------------
@@ -9652,8 +9653,7 @@ WHERE `entry` = 181056;
 -- Configure duel arbiter game object ----------------------------------------
 UPDATE `gameobject_template`
 SET
-    `faction` = 5,
-    `size` = 1
+    `faction` = 5
 WHERE `entry` = 21680;
 
 -- Insert stationery transports ----------------------------------------------