+++ _updates/world/0004_world_fishing.sql
@@ -7,46 +7,7 @@
 --
 
 -- Provide variables for game objects ----------------------------------------
-SET @GO_TYPE_DOOR                               = 0;
-SET @GO_TYPE_BUTTON                             = 1;
-SET @GO_TYPE_QUESTGIVER                         = 2;
-SET @GO_TYPE_CHEST                              = 3;
-SET @GO_TYPE_BINDER                             = 4;
-SET @GO_TYPE_GENERIC                            = 5;
-SET @GO_TYPE_TRAP                               = 6;
-SET @GO_TYPE_CHAIR                              = 7;
-SET @GO_TYPE_SPELL_FOCUS                        = 8;
-SET @GO_TYPE_TEXT                               = 9;
-SET @GO_TYPE_GOOBER                             = 10;
-SET @GO_TYPE_TRANSPORT                          = 11;
-SET @GO_TYPE_AREA_DAMAGE                        = 12;
-SET @GO_TYPE_CAMERA                             = 13;
-SET @GO_TYPE_MAP_OBJECT                         = 14;
-SET @GO_TYPE_MO_TRANSPORT                       = 15;
-SET @GO_TYPE_DUEL_ARBITER                       = 16;
 SET @GO_TYPE_FISHING_NODE                       = 17;
-SET @GO_TYPE_SUMMONING_RITUAL                   = 18;
-SET @GO_TYPE_MAILBOX                            = 19;
-SET @GO_TYPE_AUCTION_HOUSE                      = 20;
-SET @GO_TYPE_GUARD_POST                         = 21;
-SET @GO_TYPE_SPELL_CASTER                       = 22;
-SET @GO_TYPE_MEETING_STONE                      = 23;
-SET @GO_TYPE_FLAG_STAND                         = 24;
-SET @GO_TYPE_FISHING_HOLE                       = 25;
-SET @GO_TYPE_FLAG_DROP                          = 26;
-SET @GO_TYPE_MINI_GAME                          = 27;
-SET @GO_TYPE_LOTTERY_KIOSK                      = 28;
-SET @GO_TYPE_CAPTURE_POINT                      = 29;
-SET @GO_TYPE_AURA_GENERATOR                     = 30;
-
--- -> Game object flags ------------------------------------------------------
-SET @GO_FLAG_IN_USE                             = 1;
-SET @GO_FLAG_LOCKED                             = 2;
-SET @GO_FLAG_INTERACT_COND                      = 4;
-SET @GO_FLAG_TRANSPORT                          = 8;
-SET @GO_FLAG_NO_INTERACT                        = 16;
-SET @GO_FLAG_NODESPAWN                          = 32;
-SET @GO_FLAG_TRIGGERED                          = 64;
 
 -- Provide variables for various item stats ----------------------------------
 SET @ITEM_QUALITY_COMMON                        = 1;
@@ -55,6 +16,15 @@ SET @ITEM_SUBCLASS_WEAPON_FISHING_POLE          = 20;
 SET @INVTYPE_2HWEAPON                           = 17;
 SET @MATERIAL_WOOD                              = 2;
 SET @SHEATHETYPE_MAINHAND                       = 1;
+
+-- Update world database version ---------------------------------------------
+UPDATE `db_version`
+SET
+    `version`             = 'mangos-zero - version 0.1.4 (development)',
+    `creature_ai_version` = 'mangos-zero Artificial Creature Intelligence - version 0.1.4 (development)'
+WHERE
+    `version`             = 'mangos-zero - version 0.1.3 (development)'
+;
 
 -- Insert fishing bobber templates -------------------------------------------
 INSERT INTO `gameobject_template`
@@ -66,8 +36,7 @@ VALUES
 -- Configure the fishing bobber ----------------------------------------------
 UPDATE `gameobject_template`
 SET
-    `faction`   = 35, -- friendly
-    `size`      = 1
+    `faction`   = 35 -- friendly
 WHERE
     `entry` IN (35591)
 ;